#2 Added a new property "size" to Degrees of Freedom
This implicitly also requires that the DVR and the FBR have the same
size. I do not see a useful application for different sizes except
ill-conceived optimization, though, and doing so complicates all
downstream interfaces. For example, the Grid would have to offer a
dvr_shape and an fbr_shape.