#24 Cleaned up Expression interface
The time t was optional, which would generally work, but for non-obvious
reasons. As there is no thorough concept for time-independent systems
yet, and because you rarely use expressions directly, I made time mandatory for now.

Descriptions of the apply() functions were almost duplicated; improved
the base class description and removed the near-duplicates.