#11 Added and used a custom clipping function
We choose to clip only the real-valued part of the spectrum and
leave the imaginary parts untouched.

You normally truncate to improve convergence w.r.t. the timestep. The
imaginary part of the eigenvalues is usually associated with absorbing
boundary conditions, and should never dominate the dynamics in any sane
setup. So there is some freedom in the choice of the cutoff mechanism.
We choose to clip only the real values because that looks like the least
surprising behavior.